English Tutor/requirements.txt
google-generativeai
streamlit
openai-whisper 
pydub